refactor(frontend): extract NLQ logic from TopBar into dedicated AINLQ hook
- Create useAINLQ hook in common/ai/AINLQ.tsx with all NLQ orchestration
- Simplify TopBar.tsx to consume the hook via callbacks
- Remove XTM One → legacy AI fallback when XTM One is configured
- Add user-friendly, i18n-translated error messages for all NLQ failure paths
- Add "NLQ mutation failed" translation key across all 9 languages
- Replace "XTM One is unreachable, falling back to built-in AI" with "XTM One is unreachable"
