+++ opencti-platform/opencti-front/lang/front/de.json
@@ -2806,6 +2806,7 @@
   "NLQ agent returned an unreadable response": "NLQ-Agent hat eine unlesbare Antwort zurückgegeben",
   "NLQ agent returned invalid filters": "NLQ-Agent hat ungültige Filter zurückgegeben",
   "NLQ call failed": "NLQ-Aufruf fehlgeschlagen",
+  "NLQ mutation failed": "NLQ-Mutation fehlgeschlagen",
   "No": "Nein",
   "no access": "kein Zugriff",
   "No active connectors": "Keine aktiven Anschlüsse",
@@ -4803,7 +4804,7 @@
   "XTM Hub is a central forum to access resources, share tradecraft, and optimize the use of Filigran's products, fostering collaboration and empowering the community.": "Der XTM Hub ist ein zentrales Forum für den Zugang zu Ressourcen, den Austausch von Fachwissen und die Optimierung der Nutzung von Filigran-Produkten, um die Zusammenarbeit zu fördern und die Gemeinschaft zu stärken.",
   "XTM Hub is the focal point to find every Services and Products.": "XTM Hub ist der zentrale Punkt, um alle Dienste und Produkte zu finden.",
   "XTM Hub Registration": "XTM Hub-Registrierung",
-  "XTM One is unreachable, falling back to built-in AI": "XTM One ist nicht erreichbar und greift auf die integrierte KI zurück",
+  "XTM One is unreachable": "XTM One ist unerreichbar",
   "XTM_ONE_REGISTRATION_MANAGER": "XTM One Registrierungsmanager",
   "yara": "YARA",
   "Year": "Jahr",
