+++ opencti-platform/opencti-front/lang/front/en.json
@@ -2806,6 +2806,7 @@
   "NLQ agent returned an unreadable response": "NLQ agent returned an unreadable response",
   "NLQ agent returned invalid filters": "NLQ agent returned invalid filters",
   "NLQ call failed": "NLQ call failed",
+  "NLQ mutation failed": "NLQ mutation failed",
   "No": "No",
   "no access": "no access",
   "No active connectors": "No active connectors",
@@ -4803,7 +4804,7 @@
   "XTM Hub is a central forum to access resources, share tradecraft, and optimize the use of Filigran's products, fostering collaboration and empowering the community.": "XTM Hub is a central forum to access resources, share tradecraft, and optimize the use of Filigran's products, fostering collaboration and empowering the community.",
   "XTM Hub is the focal point to find every Services and Products.": "XTM Hub is the focal point to find every Services and Products.",
   "XTM Hub Registration": "XTM Hub Registration",
-  "XTM One is unreachable, falling back to built-in AI": "XTM One is unreachable, falling back to built-in AI",
+  "XTM One is unreachable": "XTM One is unreachable",
   "XTM_ONE_REGISTRATION_MANAGER": "XTM One registration manager",
   "yara": "YARA",
   "Year": "Year",
