+++ opencti-platform/opencti-front/lang/front/es.json
@@ -2806,6 +2806,7 @@
   "NLQ agent returned an unreadable response": "El agente NLQ devolvió una respuesta ilegible",
   "NLQ agent returned invalid filters": "El agente NLQ devolvió filtros no válidos",
   "NLQ call failed": "Llamada NLQ fallida",
+  "NLQ mutation failed": "Mutación NLQ fallida",
   "No": "No",
   "no access": "sin acceso",
   "No active connectors": "Sin conectores activos",
@@ -4803,7 +4804,7 @@
   "XTM Hub is a central forum to access resources, share tradecraft, and optimize the use of Filigran's products, fostering collaboration and empowering the community.": "XTM Hub es un foro central para acceder a recursos, compartir técnicas y optimizar el uso de los productos de Filigran, fomentando la colaboración y potenciando a la comunidad.",
   "XTM Hub is the focal point to find every Services and Products.": "XTM Hub es el punto focal para encontrar todos los servicios y productos.",
   "XTM Hub Registration": "Registro en el Centro XTM",
-  "XTM One is unreachable, falling back to built-in AI": "XTM uno es inalcanzable, cayendo de nuevo a AI incorporado",
+  "XTM One is unreachable": "XTM Uno es inalcanzable",
   "XTM_ONE_REGISTRATION_MANAGER": "Gestor de registro XTM One",
   "yara": "YARA",
   "Year": "Año",
