+++ opencti-platform/opencti-front/lang/front/fr.json
@@ -2806,6 +2806,7 @@
   "NLQ agent returned an unreadable response": "L'agent NLQ a renvoyé une réponse illisible",
   "NLQ agent returned invalid filters": "L'agent NLQ a renvoyé des filtres non valides",
   "NLQ call failed": "L'appel NLQ a échoué",
+  "NLQ mutation failed": "La requête NLQ a échoué",
   "No": "Non",
   "no access": "aucun accès",
   "No active connectors": "Pas de connecteurs actifs",
@@ -4803,7 +4804,7 @@
   "XTM Hub is a central forum to access resources, share tradecraft, and optimize the use of Filigran's products, fostering collaboration and empowering the community.": "XTM Hub est un forum central permettant d'accéder aux ressources, de partager les techniques et d'optimiser l'utilisation des produits Filigran, en encourageant la collaboration et en renforçant la communauté.",
   "XTM Hub is the focal point to find every Services and Products.": "XTM Hub est le point focal pour trouver tous les services et produits.",
   "XTM Hub Registration": "Enregistrement du XTM Hub",
-  "XTM One is unreachable, falling back to built-in AI": "XTM One est injoignable, utilisation de l'IA intégrée",
+  "XTM One is unreachable": "XTM One est injoignable",
   "XTM_ONE_REGISTRATION_MANAGER": "Gestionnaire d'enregistrement XTM One",
   "yara": "YARA",
   "Year": "Année",
