+++ opencti-platform/opencti-front/lang/front/it.json
@@ -2806,6 +2806,7 @@
   "NLQ agent returned an unreadable response": "L'agente NLQ ha restituito una risposta illeggibile",
   "NLQ agent returned invalid filters": "L'agente NLQ ha restituito filtri non validi",
   "NLQ call failed": "Chiamata NLQ fallita",
+  "NLQ mutation failed": "Mutazione NLQ fallita",
   "No": "No",
   "no access": "nessun accesso",
   "No active connectors": "Nessun connettore attivo",
@@ -4803,7 +4804,7 @@
   "XTM Hub is a central forum to access resources, share tradecraft, and optimize the use of Filigran's products, fostering collaboration and empowering the community.": "XTM Hub è un forum centrale per l'accesso alle risorse, la condivisione delle tecniche e l'ottimizzazione dell'uso dei prodotti Filigran'promuovendo la collaborazione e il potenziamento della comunità.",
   "XTM Hub is the focal point to find every Services and Products.": "XTM Hub è il punto focale per trovare tutti i servizi e prodotti.",
   "XTM Hub Registration": "Registrazione a XTM Hub",
-  "XTM One is unreachable, falling back to built-in AI": "XTM One è irraggiungibile, si ripiega sull'AI integrata",
+  "XTM One is unreachable": "XTM One non è raggiungibile",
   "XTM_ONE_REGISTRATION_MANAGER": "Gestore registrazione XTM One",
   "yara": "YARA",
   "Year": "Anno",
