+++ opencti-platform/opencti-front/lang/front/ja.json
@@ -2806,6 +2806,7 @@
   "NLQ agent returned an unreadable response": "NLQエージェントは読めない応答を返した",
   "NLQ agent returned invalid filters": "NLQ エージェントが無効なフィルタを返した",
   "NLQ call failed": "NLQコールに失敗",
+  "NLQ mutation failed": "NLQ変異に失敗",
   "No": "いいえ",
   "no access": "アクセスなし",
   "No active connectors": "アクティブ・コネクターなし",
@@ -4803,7 +4804,7 @@
   "XTM Hub is a central forum to access resources, share tradecraft, and optimize the use of Filigran's products, fostering collaboration and empowering the community.": "XTM Hubは、リソースにアクセスし、技術を共有し、Filigran'製品の使用を最適化するための中心的なフォーラムであり、コラボレーションを促進し、コミュニティに力を与えます。",
   "XTM Hub is the focal point to find every Services and Products.": "XTMハブは、あらゆるサービスや製品を検索できる中心的な場所です。",
   "XTM Hub Registration": "XTMハブ登録",
-  "XTM One is unreachable, falling back to built-in AI": "XTM Oneにアクセスできず、内蔵AIに戻る",
+  "XTM One is unreachable": "XTMワンが到達不能",
   "XTM_ONE_REGISTRATION_MANAGER": "XTM One登録マネージャー",
   "yara": "YARA",
   "Year": "年",
