+++ opencti-platform/opencti-front/lang/front/ko.json
@@ -2806,6 +2806,7 @@
   "NLQ agent returned an unreadable response": "NLQ 에이전트가 읽을 수 없는 응답을 반환했습니다",
   "NLQ agent returned invalid filters": "NLQ 에이전트가 잘못된 필터를 반환했습니다",
   "NLQ call failed": "NLQ 호출에 실패했습니다",
+  "NLQ mutation failed": "NLQ 돌연변이 실패",
   "No": "아니요",
   "no access": "접근 없음",
   "No active connectors": "활성 커넥터 없음",
@@ -4803,7 +4804,7 @@
   "XTM Hub is a central forum to access resources, share tradecraft, and optimize the use of Filigran's products, fostering collaboration and empowering the community.": "XTM 허브는 리소스에 액세스하고, 트레이딩 기술을 공유하고, 필리그란의 제품 사용을 최적화하여 협업을 촉진하고 커뮤니티의 역량을 강화하는 중앙 포럼입니다.",
   "XTM Hub is the focal point to find every Services and Products.": "XTM Hub는 모든 서비스와 상품을 찾을 수 있는 중심점입니다.",
   "XTM Hub Registration": "XTM 허브 등록",
-  "XTM One is unreachable, falling back to built-in AI": "XTM One에 연결할 수 없어 내장형 AI로 되돌아갑니다",
+  "XTM One is unreachable": "XTM One에 연결할 수 없습니다",
   "XTM_ONE_REGISTRATION_MANAGER": "XTM One 등록 관리자",
   "yara": "YARA",
   "Year": "년",
