+++ opencti-platform/opencti-front/lang/front/ru.json
@@ -2806,6 +2806,7 @@
   "NLQ agent returned an unreadable response": "Агент NLQ вернул нечитаемый ответ",
   "NLQ agent returned invalid filters": "Агент NLQ вернул недопустимые фильтры",
   "NLQ call failed": "Вызов NLQ не удался",
+  "NLQ mutation failed": "Мутация NLQ не удалась",
   "No": "Нет",
   "no access": "нет доступа",
   "No active connectors": "Нет активных коннекторов",
@@ -4803,7 +4804,7 @@
   "XTM Hub is a central forum to access resources, share tradecraft, and optimize the use of Filigran's products, fostering collaboration and empowering the community.": "XTM Hub - это центральный форум для доступа к ресурсам, обмена опытом и оптимизации использования продуктов Filigran', способствующий сотрудничеству и расширению возможностей сообщества.",
   "XTM Hub is the focal point to find every Services and Products.": "XTM Hub - это координационный центр для поиска всех услуг и продуктов.",
   "XTM Hub Registration": "Регистрация концентратора XTM",
-  "XTM One is unreachable, falling back to built-in AI": "XTM One недоступен, приходится обращаться к встроенному искусственному интеллекту",
+  "XTM One is unreachable": "XTM One недоступен",
   "XTM_ONE_REGISTRATION_MANAGER": "Менеджер регистрации XTM One",
   "yara": "яра",
   "Year": "Год",
