+++ opencti-platform/opencti-front/lang/front/zh.json
@@ -2806,6 +2806,7 @@
   "NLQ agent returned an unreadable response": "NLQ 代理返回无法读取的响应",
   "NLQ agent returned invalid filters": "NLQ 代理返回无效过滤器",
   "NLQ call failed": "NLQ 调用失败",
+  "NLQ mutation failed": "NLQ 突变失败",
   "No": "否",
   "no access": "无法访问",
   "No active connectors": "无活动连接器",
@@ -4803,7 +4804,7 @@
   "XTM Hub is a central forum to access resources, share tradecraft, and optimize the use of Filigran's products, fostering collaboration and empowering the community.": "XTM Hub 是一个中心论坛，用于获取资源、分享技术和优化 Filigran&apos 产品的使用，促进合作并增强社区能力。",
   "XTM Hub is the focal point to find every Services and Products.": "XTM Hub 是查找所有服务和产品的中心点。",
   "XTM Hub Registration": "XTM Hub 注册",
-  "XTM One is unreachable, falling back to built-in AI": "XTM One 无法连接，转而使用内置人工智能",
+  "XTM One is unreachable": "XTM One 无法连接",
   "XTM_ONE_REGISTRATION_MANAGER": "XTM One 注册管理器",
   "yara": "YARA",
   "Year": "年",
